refactor: decouple event form from the page component
So the form can be re-used by edit event later on
