+++ app/frontend/src/components/Datepicker.tsx
@@ -6,7 +6,7 @@ import {
 } from "@material-ui/pickers";
 import { CHANGE_DATE } from "features/constants";
 import { Control, Controller, UseControllerOptions } from "react-hook-form";
-import dayjs from "utils/dayjs";
+import dayjs, { Dayjs } from "utils/dayjs";
 
 import { dateFormats } from "./constants";
 
@@ -19,6 +19,7 @@ const getLocaleFormat = () => {
 interface DatepickerProps {
   className?: string;
   control: Control;
+  defaultValue?: Dayjs;
   error: boolean;
   helperText: React.ReactNode;
   id: string;
@@ -32,6 +33,7 @@ interface DatepickerProps {
 export default function Datepicker({
   className,
   control,
+  defaultValue,
   error,
   helperText,
   id,
@@ -45,7 +47,7 @@ export default function Datepicker({
     <MuiPickersUtilsProvider utils={DateFnsUtils}>
       <Controller
         control={control}
-        defaultValue={dayjs()}
+        defaultValue={defaultValue ?? dayjs()}
         name={name}
         rules={rules}
         render={({ onChange, value }) => (
